release 0 and release 1, rspec test still pass
